requirements.txt
# Duke Climate Risk Engine — Dependencies

# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
xarray>=2023.1.0
scipy>=1.11.0
netCDF4>=1.6.0

# Geospatial
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
fiona>=1.9.0
pyogrio>=0.7.0
pystac-client>=0.7.0
planetary-computer>=0.4.0
# pdal and laspy installed separately (complex C++ deps)
# pip install pdal or conda install -c conda-forge pdal
laspy>=2.5.0
richdem>=2.3.0    # Terrain analysis (slope, aspect, flow routing)

# Machine learning
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.0       # ViT implementations
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.43.0
einops>=0.7.0     # Tensor rearrangement for ViT

# Computer vision
Pillow>=10.0.0
albumentations>=1.3.0    # Image augmentation

# Visualization
folium>=0.15.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
branca>=0.7.0     # Folium colormap support

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
httpx>=0.25.0     # Async HTTP client for external APIs

# PDF generation
reportlab>=4.0.0
weasyprint>=60.0

# Utilities
tqdm>=4.66.0
loguru>=0.7.0
PyYAML>=6.0.1
omegaconf>=2.3.0  # Config management
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
boto3>=1.34.0     # AWS S3 for NAIP data
joblib>=1.3.0     # Parallel processing
pyarrow>=14.0.0   # Parquet I/O

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.6.0

# Colab extras (installed automatically on Colab)
# google-colab  -- pre-installed on Colab
# google-auth   -- pre-installed on Colab
